backend/requirements.txt
aiohttp==3.3.0
aiohttp-cors==0.7.0
motor==1.2.2
pymongo==3.6.1
PyYAML==3.12
